Add $api.content, with Store.map function allowing a subtree to be mapped to another Store
