+++ loader/Loader.fifty.ts
@@ -11,96 +11,6 @@ namespace slime {
 		}
 	}
 
-	export namespace runtime.content {
-		export interface Store<T> {
-			get: (path: string[]) => slime.$api.fp.Maybe<T>
-		}
-
-		export interface ContentEntry<T> {
-			name: string
-			value: T
-		}
-
-		export interface IndexEntry<T> {
-			name: string
-			index: Index<T>
-		}
-
-		export type Entry<T> = ContentEntry<T> | IndexEntry<T>
-
-		export interface Index<T> extends Store<T> {
-			list: (path: string[]) => slime.$api.fp.Maybe<Entry<T>[]>
-		}
-
-		(
-			function(
-				fifty: slime.fifty.test.Kit
-			) {
-				const { verify } = fifty;
-
-				fifty.tests.mock = function() {
-					var code: slime.runtime.test.Script = fifty.$loader.script("fixtures.ts");
-					var fixtures = code();
-					var content: slime.runtime.test.mock.Content<string> = fixtures.mock.content();
-
-					content.set("lower/a", "a");
-					content.set("lower/b", "b");
-					content.set("upper/b", "B");
-					content.set("asterisk", "*");
-
-					const nullable = <T>(m: slime.$api.fp.Maybe<T>): T => {
-						if (m.present) return m.value;
-						return null;
-					}
-
-					const withName: (name: string) => (entries: Entry<string>[]) => Entry<string> = function(name) {
-						return function(entries) {
-							return entries.find(function(entry) { return entry.name == name; });
-						}
-					};
-
-					const hasName: (name: string) => (entries: Entry<string>[]) => boolean = function(name) {
-						return function(entries) {
-							return Boolean(withName(name)(entries));
-						}
-					};
-
-					const isIndexEntry = function(i: Entry<string>): i is IndexEntry<string> {
-						return Boolean(i["index"]);
-					}
-
-					var store = content.store;
-					verify(store).get(["lower","b"]).evaluate(nullable).is("b");
-					verify(store).get(["upper","b"]).evaluate(nullable).is("B");
-					verify(store).get(["a"]).evaluate(nullable).is(null);
-
-					var index = content.index;
-					var top = nullable(index.list([]));
-					verify(top).length.is(3);
-					verify(top).evaluate(hasName("lower")).is(true);
-					verify(top).evaluate(hasName("upper")).is(true);
-					verify(top).evaluate(hasName("asterisk")).is(true);
-					verify(top).evaluate(hasName("aaa")).is(false);
-					verify(top).evaluate(withName("lower")).evaluate(isIndexEntry).is(true);
-					verify(top).evaluate(withName("upper")).evaluate(isIndexEntry).is(true);
-					verify(top).evaluate(withName("asterisk")).evaluate(isIndexEntry).is(false);
-
-					var lower = nullable(index.list(["lower"]));
-					verify(lower).length.is(2);
-					verify(lower).evaluate(hasName("a")).is(true);
-					verify(lower).evaluate(hasName("b")).is(true);
-
-					var upper = nullable(index.list(["upper"]));
-					verify(upper).length.is(1);
-
-					var no = nullable(index.list(["foo"]));
-					verify(no).is(null);
-				}
-			}
-		//@ts-ignore
-		)(fifty);
-	}
-
 	export namespace runtime.loader {
 		export interface Node {
 			name: string
@@ -373,7 +283,6 @@ namespace slime.runtime.internal.loader {
 			fifty: slime.fifty.test.Kit
 		) {
 			fifty.tests.suite = function() {
-				fifty.run(fifty.tests.mock);
 			}
 		}
 	//@ts-ignore
